% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultigeneSurvival.R
\name{MultigeneSurvival}
\alias{MultigeneSurvival}
\title{Multiple Gene Survival Analysis}
\usage{
MultigeneSurvival(dataset, time, event, title, alpha)
}
\arguments{
\item{dataset}{A dataset. This dataset contains only gene expression and survival information.}

\item{time}{A character. The time variable name in dataset.}

\item{event}{A character. The event variable name in dataset.}

\item{title}{A character. Usually it's data set name.}

\item{alpha}{A number. alpha = 1 is LASSO, 0 < alpha < 1 is Elastic net.}
}
\description{
This function used LASSO/Elastic net in cox regreesion to determine the important genes and used Kmeans (K=2) to divide them into 2 groups in survival.
The output includes K-M plot, a list of important genes selected from LASSO/Elastic in Cox model.
}
\details{
1. For each gene, the gene expression will be standardized to be mean equal to 0 and standard deviation equal to 1. If we don’t do so, we couldn’t compare the coefficient estimates across genes.

2. Use LASSO/Elastic method for variables selection in the Cox model. The predictor is gene and the outcome is survival. For those coefficient converging to 0 in Cox model are not related to survival, while the remaining nonzero coefficients are chosen for the next step.

3.	To use LASSO/Elastic, we have to determine tuning parameter λ. The cross validation for 100 times was used to determine the most appropriate λ so that the cross validation error is minimized.

4.	After choosing important genes, i.e. coefficients are not equal to 0, we will use standardized gene expression to separate subjects into two groups based on k-means (k=2 here).

5.	Plot the Kaplan-Meier curve from the k-means cluster result.
}
\examples{
\dontrun{
data(chin_expr_surv)
MultigeneSurvival(dataset=chin_expr_surv,time="time",event="event",title="Chin",alpha=1)
}

}
